# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de renta de herramientas con aceptación mutua (apretón de manos digital).\n\n## Flujo de renta\n1. Propietario registra herramienta con precio\n2. Solicitante solicita renta → estado **pending**\n3. Ambos llaman `/confirm-delivery` → estado **active**\n4. Ambos llaman `/confirm-return` → estado **completed** (fondos liberados)",
        "title": "Tool Rental API",
        "contact": {
            "name": "Angel Chame",
            "email": "angelchame6@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/rentals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todas las rentas donde el usuario es propietario o solicitante",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rentas"
                ],
                "summary": "Mis rentas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RentalResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "El solicitante crea una solicitud de renta. La herramienta queda reservada (estado: pending)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rentas"
                ],
                "summary": "Solicitar renta",
                "parameters": [
                    {
                        "description": "Datos de la renta",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRentalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RentalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Herramienta no encontrada",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "409": {
                        "description": "Herramienta no disponible",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/rentals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve el estado completo de una renta, incluyendo el estado del apretón de manos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rentas"
                ],
                "summary": "Detalle de renta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la renta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RentalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancela una renta pendiente o activa. Libera la herramienta automáticamente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rentas"
                ],
                "summary": "Cancelar renta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la renta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RentalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "409": {
                        "description": "No se puede cancelar una renta completada",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/rentals/{id}/confirm-return": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paso 3: propietario Y solicitante deben confirmar la devolución física. Cuando ambos confirman → estado pasa a \"completed\" y se liberan los fondos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rentas"
                ],
                "summary": "Confirmar devolución (apretón de manos)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la renta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RentalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "409": {
                        "description": "Ya confirmaste o la renta no está activa",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateRentalRequest": {
            "type": "object",
            "required": [
                "end_date",
                "start_date",
                "tool_id"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2026-06-25T18:00:00Z"
                },
                "start_date": {
                    "type": "string",
                    "example": "2026-06-20T09:00:00Z"
                },
                "tool_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "dto.ErrResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "mensaje de error"
                }
            }
        },
...
        "dto.RentalResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_rate": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner_confirmed_delivery": {
                    "description": "Estado del apretón de manos en entrega",
                    "type": "boolean"
                },
                "owner_confirmed_return": {
                    "type": "boolean"
                },
                "owner_id": {
                    "type": "string"
                },
                "requester_confirmed_delivery": {
                    "type": "boolean"
                },
                "requester_confirmed_return": {
                    "description": "Estado del apretón de manos en devolución",
                    "type": "boolean"
                },
                "requester_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tool_id": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Pegar el token así: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}